🔀 :: (#99) 모든 카테고리 API

+++ src/main/java/com/meturial/domain/recipe/domain/repository/CustomChoiceRecipeRepositoryImpl.java
@@ -27,7 +27,7 @@ public class CustomChoiceRecipeRepositoryImpl implements CustomChoiceRecipeRepos
                                 choiceRecipe.recipe.id,
                                 recipe.name,
                                 review.starRating,
-                                review.recipe.id.count(),
+                                review.recipe.id.count().intValue(),
                                 recipe.foodImageUrl,
                                 recipe.category
                         ))
