+++ src/main/java/com/meturial/domain/recipe/domain/repository/vo/QueryRecipeRankingVo.java
@@ -10,13 +10,13 @@ public class QueryRecipeRankingVo {
     private final UUID recipeId;
     private final String name;
     private final Float starRating;
-    private final Long starCount;
+    private final Integer starCount;
     private final String recipeImageUrl;
     private final String recipeCategory;
     private final String recipeMaterial;
 
     @QueryProjection
-    public QueryRecipeRankingVo(UUID recipeId, String name, Float starRating, Long starCount,
+    public QueryRecipeRankingVo(UUID recipeId, String name, Float starRating, Integer starCount,
                                 String recipeImageUrl, String recipeCategory, String recipeMaterial) {
         this.recipeId = recipeId;
         this.name = name;
